# simpler & faster UOp key