@roelofvandijk 2024-08-28
@Qazalin If the risk for collisions is so high here, should the hash use [here](https://github.com/tinygrad/tinygrad/blob/85591bd1ae6ca9ff1e78c999468ee9902ca0093b/tinygrad/codegen/kernel.py#L770) also become a `.key` use?

It might be useful to remove the hash function of the class to avoid calling it inadvertently.

```
  def __hash__(self):
    raise NotImplementedError
```